continue reorganization of graph package
